Fixed a couple of compiler warnings.
git-svn-id: svn://scribus.net/trunk/Scribus@11229 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/latexeditor.cpp
@@ -697,7 +697,7 @@ QIcon IconBuffer::readData()
 	QByteArray data = file->read(len);
 	QPixmap pixmap;
 	pixmap.loadFromData(data);
-	basePos += ceil(len/512.0) * 512;
+	basePos += static_cast<int>(ceil(len/512.0) * 512);
 	return QIcon(pixmap);
 }
